% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableSelection-package.R
\docType{package}
\name{coxkll-package}
\alias{coxkll}
\alias{coxkll-package}
\title{VariableSelection package}
\description{
Description
}
\seealso{
Useful links:
\itemize{
  \item \url{}
  \item Report bugs at \url{}
}

}
\author{
\strong{Maintainer}: Lingfeng Luo \email{lfluo@umich.edu}

Authors:
\itemize{
  \item Di Wang \email{yubos@umich.edu}
  \item Kevin He \email{kevinhe@umich.edu}
}

}
