% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setupLambdaCoxKL}
\alias{setupLambdaCoxKL}
\title{Setup Lambda Sequence for Cox–KL Model (Internal)}
\usage{
setupLambdaCoxKL(
  Z,
  time,
  delta,
  delta_tilde,
  RS,
  beta.init,
  stratum,
  group,
  group.multiplier,
  n.each_stratum,
  alpha,
  eta,
  nlambda,
  lambda.min.ratio
)
}
\description{
Generates a sequence of penalty parameters (\code{lambda}) and initializes coefficients
for the Cox–KL model. The maximum lambda (\code{lambda.max}) is defined as the
smallest value that shrinks all coefficients (\code{beta}) to zero. Note that different
values of \code{eta} will produce different \code{lambda} sequences.
}
